crypto: move ledger code to its own subfolder
