+++ crypto/ledger/ledger_notavail.go
@@ -1,7 +1,7 @@
 // +build !cgo !ledger
 // test_ledger_mock
 
-package crypto
+package ledger
 
 import (
 	"github.com/pkg/errors"
